Extract directives to their own file
